I modified the variable 'overdue_cost' as well as the method such that the method returns a type double, which is the new type of the variable

+++ ID.java
@@ -2,7 +2,7 @@ public class ID{
      private String name;
      private String birthday;
      private String year;//Class rank has been named 'year', as opposed to 'class' to avoid confusion with the code
-     private int overdue_cost;
+     private double overdue_cost;
 
      public boolean taken;
      
@@ -23,10 +23,10 @@ public class ID{
 	 System.out.println(year);
      }
 
-     private int calculateFee(int minutes){
+     private double calculateFee(int minutes){
       /*Flat rate fee of $1.50 will be applied for each minute gone over */
 
-	int fee = 1.5 * minutes;
+	double fee = 1.5 * minutes;
 	overdue_cost = fee;
 	return fee;
      }
